refactor(api): move `Unmarshaller` interface and `UnmarshalParams` into api
Also removed unneeded `UnmarshalFromRequest` methods
