+++ lib/fsi.go
@@ -3,7 +3,6 @@ package lib
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"os"
 	"path"
 	"path/filepath"
@@ -46,13 +45,6 @@ func (m FSIMethods) Attributes() map[string]AttributeSet {
 type LinkParams struct {
 	Dir    string `qri:"fspath"`
 	Refstr string `json:"refstr"`
-}
-
-// UnmarshalFromRequest implements a custom deserialization-from-HTTP request
-func (p *LinkParams) UnmarshalFromRequest(r *http.Request) error {
-	p.Refstr = r.FormValue("refstr")
-	p.Dir = r.FormValue("dir")
-	return nil
 }
 
 // FSIWriteParams encapsultes arguments for writing to an FSI-linked directory
